Startup database config
completed first draft of code which checks database location at startup.
This is important the first time the extension is started. Then the user
can choose a location to save the default database, and this new database
will be used whenever the extension is started.

The code is probably a bit dodgy - should test to find any problems.